requirements.txt
streamlit>=1.40.0
feedparser
beautifulsoup4
textblob
pandas
regex
requests
openai
python-dotenv
python-dateutil
